requirements.txt
Django==2.2.24
environs==5.2.1
dj-database-url==0.5.0
marshmallow==3.0.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.0